src/test/java/com/rhkr8521/pingpong/Controllers/StartRoomControllerTest.java
package com.rhkr8521.pingpong.Controllers;

import com.rhkr8521.pingpong.Services.StartRoomService;

import com.fasterxml.jackson.databind.ObjectMapper;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;
import org.springframework.http.MediaType;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.setup.MockMvcBuilders;
import java.util.HashMap;
import java.util.Map;

import static org.mockito.Mockito.verify;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.put;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;

@ExtendWith(MockitoExtension.class)
public class StartRoomControllerTest {

    @Mock
    private StartRoomService startRoomService;

    @InjectMocks
    private StartRoomController startRoomController;

    private MockMvc mockMvc;

    @BeforeEach
    public void setUp() {
        mockMvc = MockMvcBuilders.standaloneSetup(startRoomController).build();
    }

    @Test
    void testStartRoomSuccess() throws Exception {
        Integer roomId = 1;
        Map<String, Integer> requestBody = new HashMap<>();
        requestBody.put("userId", 1);

        mockMvc.perform(put("/room/start/" + roomId)
                        .contentType(MediaType.APPLICATION_JSON)
                        .content(new ObjectMapper().writeValueAsString(requestBody)))
                .andExpect(status().isOk())
                .andExpect(jsonPath("$.code").value(200))
                .andExpect(jsonPath("$.message").value("API 요청이 성공했습니다."));

        verify(startRoomService).startRoom(roomId, 1);
    }

}
